test: split azure assistants tests
