+++ specs/GH727/product.md
@@ -11,32 +11,33 @@ GH-727 / #727
 
 ## 本 tranche 目标
 
-- 拆分 `src/core/providers/thinking/tests.rs`，它当前 1100 行，是 #727 当前 top offenders 之一。
-- 按 provider group 将测试移动到独立模块，保持测试语义和 public API 不变。
-- 所有新增 Rust 文件低于 800 行。
+- 拆分 `src/core/providers/azure/assistants.rs`，它当前 1256 行，是 #727 当前 top offenders 之一。
+- 按现有 Azure `chat.rs` / `chat_tests.rs` 模式，把 inline tests 移动到 `assistants_tests.rs`。
+- 保持 Azure Assistants runtime 代码、测试断言、fixtures 和 public API 不变。
+- 所有新增或修改后的 Rust 文件低于 800 行。
 
 ## 非目标
 
-- 不修改 thinking provider runtime 代码。
-- 不重构 OpenAI/Anthropic/DeepSeek/Gemini/OpenRouter thinking 行为。
-- 不在本 PR 中处理其余 50 个大文件。
+- 不修改 Azure Assistants runtime 行为。
+- 不重构 Azure Assistants API surface、request/response types、URL building logic 或 validation rules。
+- 不在本 PR 中处理其余大文件。
 - 不关闭 #727，除非 issue owner 决定一个 tranche PR 足以满足 tracker。
 
 ## Behavior Invariants
 
-1. 所有原有 thinking tests 仍由 `core::providers::thinking::tests::*` 测试树运行。
-2. 测试移动只能改变 module path，不改变断言、fixtures 或 production code。
-3. `src/core/providers/thinking/tests/*.rs` 单文件必须低于 800 行。
-4. `cargo test core::providers::thinking --lib` 必须通过。
+1. 所有原有 Azure Assistants tests 仍由 `core::providers::azure::assistants::tests::*` 测试树运行。
+2. 测试移动只能改变 module file location，不改变断言、fixtures 或 production code。
+3. `src/core/providers/azure/assistants.rs` 和 `src/core/providers/azure/assistants_tests.rs` 必须低于 800 行。
+4. `cargo test core::providers::azure::assistants --lib --all-features` 必须通过。
 
 ## 验收标准
 
-- [ ] `src/core/providers/thinking/tests.rs` 被替换为 `tests/mod.rs` + provider-specific test files。
-- [ ] 最大新增 test file 低于 800 行。
-- [ ] Focused thinking tests 通过。
+- [ ] `src/core/providers/azure/assistants.rs` 使用 `#[path = "assistants_tests.rs"] mod tests;` 引入测试。
+- [ ] `src/core/providers/azure/assistants_tests.rs` 包含原 inline test module body，且低于 800 行。
+- [ ] Focused Azure Assistants tests 通过。
 - [ ] `cargo fmt --all -- --check` 和 `cargo check --all-features --locked` 通过。
-- [ ] PR body 明确该 PR 是 #727 的 first tranche，不自动关闭 tracker issue。
+- [ ] PR body 明确该 PR 是 #727 的 next tranche，不自动关闭 tracker issue。
 
 ## 发布说明
 
-No runtime behavior change. This is a test layout maintenance split for U-16 compliance.
+No runtime behavior change. This is an Azure Assistants test layout maintenance split for U-16 compliance.
